+++ src/Utilites/Ssr/SearchAdapter.php
@@ -8,6 +8,7 @@ use Omikron\FactFinder\Communication\Client\ClientBuilder;
 use Omikron\FactFinder\Communication\Client\ClientException;
 use Omikron\FactFinder\Shopware6\Config\Communication;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\RouterInterface;
 
@@ -17,17 +18,20 @@ class SearchAdapter
     private Communication $config;
     private PriceFormatter $priceFormatter;
     private RouterInterface $router;
+    private LoggerInterface $factfinderLogger;
 
     public function __construct(
         ClientBuilder $clientBuilder,
         Communication $config,
         PriceFormatter $priceFormatter,
         RouterInterface $router,
+        LoggerInterface $factfinderLogger,
     ) {
-        $this->clientBuilder  = $clientBuilder;
-        $this->config         = $config;
-        $this->priceFormatter = $priceFormatter;
-        $this->router         = $router;
+        $this->clientBuilder    = $clientBuilder;
+        $this->config           = $config;
+        $this->priceFormatter   = $priceFormatter;
+        $this->router           = $router;
+        $this->factfinderLogger = $factfinderLogger;
     }
 
     public function search(
@@ -35,17 +39,21 @@ class SearchAdapter
         bool $navigationRequest,
         string $salesChannelId,
     ): array {
-        $client = $this->clientBuilder
-            ->withServerUrl($this->getServerUrl())
-            ->withApiKey($this->config->getApiKey())
-            ->withVersion($this->config->getVersion())
-            ->build();
+        try {
+            $client = $this->clientBuilder
+                ->withServerUrl($this->getServerUrl())
+                ->withApiKey($this->config->getApiKey())
+                ->withVersion($this->config->getVersion())
+                ->build();
 
-        $endpoint = $this->createEndpoint($paramString, $navigationRequest, $salesChannelId);
-        $response = $client->request('GET', $endpoint);
+            $endpoint = $this->createEndpoint($paramString, $navigationRequest, $salesChannelId);
+            $response = $client->request('GET', $endpoint);
+        } catch (ClientException $e) {
+            $this->factfinderLogger->error($e->getMessage());
+        }
 
-        if (!$response) {
-            throw new ClientException('The response was empty');
+        if (empty($response)) {
+            throw new ClientException('The response was empty or not exist. Probably 400 error during the request');
         }
 
         return $this->priceFormatter->format($this->searchResult($response));