+++ spec/Storefront/Controller/ResultControllerSpec.php
@@ -63,6 +63,7 @@ class ResultControllerSpec extends ObjectBehavior
         $this->factfinderLogger         = $factfinderLogger;
         $this->beConstructedWith($config, $pageLoader, $factfinderLogger);
         $requestStack->getCurrentRequest()->willReturn($request);
+        $salesChannelContext->getSalesChannelId()->willReturn('main_sales_channel');
         $container->get('request_stack')->willReturn($requestStack);
         $this->request->attributes = $attributes;
         $attributes->get(PlatformRequest::ATTRIBUTE_SALES_CHANNEL_CONTEXT_OBJECT)->willReturn($salesChannelContext);
@@ -85,6 +86,7 @@ class ResultControllerSpec extends ObjectBehavior
         $content = 'original content';
         $this->pageLoader->load($this->request, $this->salesChannelContext)->willReturn($page);
         $this->config->isSsrActive()->willReturn(false);
+        $this->config->disableFFWebc('main_sales_channel')->willReturn(false);
         $this->twig->render('@OmikronFactFinder/storefront/page/factfinder/result.html.twig', Argument::any())->willReturn($content);
         $this->seoUrlPlaceholderHandler->replace($content, 'https://shop.com', $this->salesChannelContext)->willReturn($content);
         $response = $this->result(